@@ -1,13 +1,10 @@
-import express = require('express');
-import cors = require('cors');
-import bodyParser = require('body-parser');
+import bodyParser from 'body-parser';
+import cors from 'cors';
+import express from 'express';
 import { closeDatabase, initializeDatabase, testConnection } from './config/database';
-import canvasRoutes from './routes/canvas.routes';
-<<<<<<< HEAD
-=======
-import usersRoutes from './routes/users.routes';
 import authRoutes from './routes/auth.routes';
->>>>>>> signup-dev
+import canvasRoutes from './routes/canvas.routes';
+import usersRoutes from './routes/users.routes';
 import { DatabaseService } from './services/database.service';
 
 const app = express();
@@ -21,16 +18,16 @@ app.use(bodyParser.urlencoded({ extended: true, limit: '50mb' }));
 app.get('/api/health', async (req: express.Request, res: express.Response) => {
   try {
     const dbHealth = await DatabaseService.healthCheck();
-    res.json({ 
-      status: 'OK', 
-      message: 'Backend server is running', 
+    res.json({
+      status: 'OK',
+      message: 'Backend server is running',
       timestamp: new Date().toISOString(),
       database: dbHealth
     });
   } catch (error) {
-    res.status(500).json({ 
-      status: 'ERROR', 
-      message: 'Backend server health check failed', 
+    res.status(500).json({
+      status: 'ERROR',
+      message: 'Backend server health check failed',
       timestamp: new Date().toISOString(),
       error: error instanceof Error ? error.message : 'Unknown error'
     });
@@ -38,13 +35,9 @@ app.get('/api/health', async (req: express.Request, res: express.Response) => {
 });
 
 // API Routes
-<<<<<<< HEAD
 app.use('/api/canvas', canvasRoutes);
-=======
 app.use('/api/auth', authRoutes);
-app.use('/api/canvas', canvasRoutes);
 app.use('/api/users', usersRoutes);
->>>>>>> signup-dev
 
 // Default route
 app.get('/', (req: express.Request, res: express.Response) => {
@@ -53,23 +46,18 @@ app.get('/', (req: express.Request, res: express.Response) => {
     version: '1.0.0',
     endpoints: {
       health: '/api/health',
-<<<<<<< HEAD
-=======
       auth: {
         register: 'POST /api/auth/register',
         login: 'POST /api/auth/login',
         logout: 'POST /api/auth/logout',
         me: 'GET /api/auth/me'
       },
->>>>>>> signup-dev
       canvas: {
         save: 'POST /api/canvas/save',
         list: 'GET /api/canvas/list',
         get: 'GET /api/canvas/:id',
         update: 'PUT /api/canvas/:id',
         delete: 'DELETE /api/canvas/:id'
-<<<<<<< HEAD
-=======
       },
       users: {
         list: 'GET /api/users',
@@ -78,7 +66,6 @@ app.get('/', (req: express.Request, res: express.Response) => {
         update: 'PUT /api/users/:id',
         delete: 'DELETE /api/users/:id',
         updateLogin: 'PATCH /api/users/:id/login'
->>>>>>> signup-dev
       }
     }
   });
@@ -106,18 +93,18 @@ app.use('*', (req: express.Request, res: express.Response) => {
 // Start server function
 async function startServer() {
   try {
-    const port = process.env['PORT'] || 3001;
-    
+    const port = process.env.PORT || 3001;
+
     // Test database connection
     const isConnected = await testConnection();
     if (!isConnected) {
       console.error('❌ Failed to connect to database. Please check your XAMPP MySQL server.');
       process.exit(1);
     }
-    
+
     // Initialize database tables
     await initializeDatabase();
-    
+
     // Start the server
     app.listen(port, () => {
       console.log(`🚀 RFM Backend API server listening on http://localhost:${port}`);
@@ -125,20 +112,17 @@ async function startServer() {
       console.log(`🎨 Canvas API: Ready for operations`);
       console.log(`📋 API Documentation: http://localhost:${port}`);
     });
-    
+
     // Graceful shutdown
-    process.on('SIGINT', async () => {
+    const shutdown = async () => {
       console.log('\n🛑 Shutting down backend server...');
       await closeDatabase();
       process.exit(0);
-    });
-    
-    process.on('SIGTERM', async () => {
-      console.log('\n🛑 Shutting down backend server...');
-      await closeDatabase();
-      process.exit(0);
-    });
-    
+    };
+
+    process.on('SIGINT', shutdown);
+    process.on('SIGTERM', shutdown);
+
   } catch (error) {
     console.error('❌ Failed to start backend server:', error);
     process.exit(1);
