@@ -1,68 +1,63 @@
-// backend/index.js
 const express = require('express');
-<<<<<<< HEAD
-=======
-const cors = require('cors');
->>>>>>> signup-dev
-const pool = require('./db');
+const cors = require('cors'); // Required for frontend connection
+const pool = require('./db'); // Assumed to export the MySQL connection pool
+const bcrypt = require('bcrypt');
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-<<<<<<< HEAD
-=======
+// --- Middlewares ---
+
 // Enable CORS for Angular frontend (allow any localhost port)
 app.use(cors({
-    origin: function(origin, callback) {
-        // Allow requests with no origin (like mobile apps or curl requests)
+    origin: function (origin, callback) {
+        // Allow requests with no origin (like mobile apps)
         if (!origin) return callback(null, true);
-        
-        // Allow all localhost origins
+
+        // Allow all localhost origins (e.g., http://localhost:4200)
         if (origin.startsWith('http://localhost:') || origin.startsWith('http://127.0.0.1:')) {
             return callback(null, true);
         }
-        
+
+        // Production or blocked origin
         callback(new Error('Not allowed by CORS'));
     },
     credentials: true
 }));
 
->>>>>>> signup-dev
 // Middleware to parse JSON request bodies
 app.use(express.json());
 
-// Test DB connection
+// --- Routes ---
+
+// 1. Health/DB Test Check
 app.get('/', async (req, res) => {
     try {
         const [rows] = await pool.query('SELECT NOW() AS now');
         res.send(`DB Connected! Server time: ${rows[0].now}`);
     } catch (err) {
-        console.error(err);
+        console.error("Database Connection Check Failed:", err);
         res.status(500).send('Database connection failed');
     }
 });
 
-// ===== USERS CRUD =====
+// 2. Authentication Routes (Login)
+// Assumed to handle POST /api/auth/login and utilize auth.service.js
+const authRoutes = require('./src/routes/auth.routes');
+app.use('/api/auth', authRoutes);
+
+// 3. USERS (Employee/Admin) CRUD Routes
 
 // GET all users
-<<<<<<< HEAD
-app.get('/users', async (req, res) => {
-    try {
-        const [rows] = await pool.query('SELECT * FROM Users');
-        res.json(rows);
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ error: 'Failed to fetch users' });
-=======
 app.get('/api/users', async (req, res) => {
     try {
         const [rows] = await pool.query('SELECT * FROM Users');
-        
-        // Transform database rows to match frontend interface
+
+        // Transform database rows to match frontend interface (splitting FullName)
         const users = rows.map(row => {
             const nameParts = row.FullName ? row.FullName.split(' ') : ['', ''];
             const firstName = nameParts[0] || '';
             const lastName = nameParts.slice(1).join(' ') || '';
-            
+
             return {
                 id: row.UserId,
                 firstName: firstName,
@@ -77,7 +72,7 @@ app.get('/api/users', async (req, res) => {
                 updated_at: row.updated_at
             };
         });
-        
+
         res.json({
             success: true,
             data: users,
@@ -90,22 +85,10 @@ app.get('/api/users', async (req, res) => {
             error: 'Failed to fetch users',
             message: err.message
         });
->>>>>>> signup-dev
     }
 });
 
 // GET single user by id
-<<<<<<< HEAD
-app.get('/users/:id', async (req, res) => {
-    try {
-        const { id } = req.params;
-        const [rows] = await pool.query('SELECT * FROM Users WHERE id = ?', [id]);
-        if (rows.length === 0) return res.status(404).json({ error: 'User not found' });
-        res.json(rows[0]);
-    } catch (err) {
-        console.error(err);
-        res.status(500).json({ error: 'Failed to fetch user' });
-=======
 app.get('/api/users/:id', async (req, res) => {
     try {
         const { id } = req.params;
@@ -116,7 +99,7 @@ app.get('/api/users/:id', async (req, res) => {
                 error: 'User not found'
             });
         }
-        
+
         const row = rows[0];
         const nameParts = row.FullName ? row.FullName.split(' ') : ['', ''];
         const user = {
@@ -132,7 +115,7 @@ app.get('/api/users/:id', async (req, res) => {
             created_at: row.created_at,
             updated_at: row.updated_at
         };
-        
+
         res.json({
             success: true,
             data: user
@@ -144,59 +127,96 @@ app.get('/api/users/:id', async (req, res) => {
             error: 'Failed to fetch user',
             message: err.message
         });
->>>>>>> signup-dev
     }
 });
 
 // POST create new user
-<<<<<<< HEAD
-app.post('/users', async (req, res) => {
-=======
 app.post('/api/users', async (req, res) => {
->>>>>>> signup-dev
     try {
-        const { username, email, password } = req.body; // adjust columns as per your table
+        const { FullName, Email, Password } = req.body;
+        // IMPORTANT: Hash the password before insertion in a real app!
+        // For this example, we assume Password is the raw value, but DB expects PasswordHash
+        const hashedPassword = await bcrypt.hash(Password, 10);
+
         const [result] = await pool.query(
-            'INSERT INTO Users (username, email, password) VALUES (?, ?, ?)',
-            [username, email, password]
+            'INSERT INTO Users (FullName, Email, PasswordHash) VALUES (?, ?, ?)',
+            [FullName, Email, hashedPassword]
         );
-        res.status(201).json({ id: result.insertId, username, email });
+
+        const newUserId = result.insertId;
+
+        res.status(201).json({
+            success: true,
+            data: {
+                id: newUserId,
+                fullName: FullName,
+                email: Email
+            },
+            message: 'User created successfully.'
+        });
     } catch (err) {
         console.error(err);
-        res.status(500).json({ error: 'Failed to create user' });
+        res.status(500).json({
+            success: false,
+            error: 'Failed to create user',
+            message: err.message
+        });
     }
 });
 
 // PUT update user
-<<<<<<< HEAD
-app.put('/users/:id', async (req, res) => {
-=======
 app.put('/api/users/:id', async (req, res) => {
->>>>>>> signup-dev
     try {
         const { id } = req.params;
-        const { username, email, password } = req.body;
-        const [result] = await pool.query(
-            'UPDATE Users SET username = ?, email = ?, password = ? WHERE id = ?',
-            [username, email, password, id]
-        );
-        if (result.affectedRows === 0) return res.status(404).json({ error: 'User not found' });
-        res.json({ id, username, email });
+        const { FullName, Email, Password, Phone, Roles, Status } = req.body;
+
+        let updateFields = [];
+        let updateValues = [];
+
+        if (FullName) { updateFields.push('FullName = ?'); updateValues.push(FullName); }
+        if (Email) { updateFields.push('Email = ?'); updateValues.push(Email); }
+        if (Phone) { updateFields.push('Phone = ?'); updateValues.push(Phone); }
+        if (Roles) { updateFields.push('Roles = ?'); updateValues.push(Roles); }
+        if (Status) { updateFields.push('Status = ?'); updateValues.push(Status); }
+
+        if (Password) {
+            const hashedPassword = await bcrypt.hash(Password, 10);
+            updateFields.push('PasswordHash = ?');
+            updateValues.push(hashedPassword);
+        }
+
+        if (updateFields.length === 0) {
+            return res.status(400).json({ success: false, error: 'No fields to update' });
+        }
+
+        const sql = `UPDATE Users SET ${updateFields.join(', ')} WHERE UserId = ?`;
+        updateValues.push(id);
+
+        const [result] = await pool.query(sql, updateValues);
+
+        if (result.affectedRows === 0) {
+            return res.status(404).json({ success: false, error: 'User not found' });
+        }
+
+        res.json({
+            success: true,
+            message: 'User updated successfully'
+        });
     } catch (err) {
         console.error(err);
-        res.status(500).json({ error: 'Failed to update user' });
+        res.status(500).json({
+            success: false,
+            error: 'Failed to update user',
+            message: err.message
+        });
     }
 });
 
 // DELETE user
-<<<<<<< HEAD
-app.delete('/users/:id', async (req, res) => {
-=======
 app.delete('/api/users/:id', async (req, res) => {
->>>>>>> signup-dev
     try {
         const { id } = req.params;
-        const [result] = await pool.query('DELETE FROM Users WHERE id = ?', [id]);
+        const [result] = await pool.query('DELETE FROM Users WHERE UserId = ?', [id]);
         if (result.affectedRows === 0) return res.status(404).json({ error: 'User not found' });
         res.json({ message: 'User deleted successfully' });
     } catch (err) {
